Procedural finished

+++ RogueLike/Assets/Scripts/Control/Map/RoomConfig.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -6,14 +7,14 @@ public class RoomConfig
 {
     /*Esta classe serve para configurar itens de drop, portas de saida, id da room e mais*/
     private int idRoom;
-    private int directionDoors; // 0000 -> cima, direita, baixo, esquerda
+    private Byte directionDoors; // 0000 -> cima, direita, baixo, esquerda
     private bool isChest;
 
     public int IdRoom{
         get => idRoom;
         set => idRoom = value;
     } 
-    public int DirectionDoors{
+    public Byte DirectionDoors{
         get => directionDoors;
         set => directionDoors = value;
     }
